implement tests for Expand Query Option
